INITIAL: iris for lookup
Nothing is working yet, still striving to have a settled style of code
for the project so that when this become bigger, it doesn't become a
huge mess as well.

A few things to consider here, no library for parsing the HTML code is
chosen yet, that will take some time.

As you will see, there is not library to parse options, we're doing it
built-in as that will be more efficient and performant and we create
more and more options.

Compatibility layers are yet not considered, but as we remain to simple
C++ code and libraries such as cURL, we can make this work on various
platforms.

Speed of downloading a WEB page has yet to be optimized.

See TODO.txt for what I consider now the priority things to work on.